Split subschemas into separate files
