+++ core/src/main/scala-3/caliban/schema/DerivationUtils.scala
@@ -1,112 +1,10 @@
 package caliban.schema
 
-import caliban.Value.EnumValue
 import caliban.introspection.adt.*
 import caliban.parsing.adt.Directive
 import caliban.schema.Annotations.*
-import caliban.schema.DerivationUtils.*
-import caliban.schema.Step.{ MetadataFunctionStep, ObjectStep }
 import caliban.schema.Types.*
 import magnolia1.TypeInfo
-
-final private class ValueTypeSchema[R, A](
-  _schema: => Schema[R, Any],
-  info: TypeInfo,
-  anns: List[Any]
-) extends Schema[R, A] {
-  private val name = getName(anns, info)
-
-  private lazy val schema = _schema
-
-  def toType(isInput: Boolean, isSubscription: Boolean): __Type =
-    if (anns.contains(GQLValueType(true))) makeScalar(name, getDescription(anns))
-    else schema.toType_(isInput, isSubscription)
-
-  def resolve(value: A): Step[R] = schema.resolve(value.asInstanceOf[Product].productElement(0))
-}
-
-final private class EnumValueSchema[R, A](
-  info: TypeInfo,
-  anns: List[Any]
-) extends Schema[R, A] {
-
-  def toType(isInput: Boolean, isSubscription: Boolean): __Type =
-    if (isInput) mkInputObject[R](anns, Nil, info)(isInput, isSubscription)
-    else mkObject[R](anns, Nil, info)(isInput, isSubscription)
-
-  private val step               = PureStep(EnumValue(getName(anns, info)))
-  def resolve(value: A): Step[R] = step
-}
-
-final private class ObjectSchema[R, A](
-  _fields: => List[(String, Schema[R, Any], Int)],
-  info: TypeInfo,
-  anns: List[Any],
-  paramAnnotations: Map[String, List[Any]]
-) extends Schema[R, A] {
-  private val name = getName(anns, info)
-
-  private lazy val fields = _fields.map { (label, schema, index) =>
-    val fieldAnnotations = paramAnnotations.getOrElse(label, Nil)
-    (getName(fieldAnnotations, label), fieldAnnotations, schema, index)
-  }
-
-  def toType(isInput: Boolean, isSubscription: Boolean): __Type =
-    if (isInput) mkInputObject[R](anns, fields, info)(isInput, isSubscription)
-    else mkObject[R](anns, fields, info)(isInput, isSubscription)
-
-  def resolve(value: A): Step[R] = MetadataFunctionStep[R] { f =>
-    val fb = Map.newBuilder[String, Step[R]]
-
-    var remaining = fields
-    while (!remaining.isEmpty) {
-      val (name, _, schema, i) = remaining.head
-      if (f.fieldNames.contains(name)) fb += name -> schema.resolve(value.asInstanceOf[Product].productElement(i))
-      remaining = remaining.tail
-    }
-
-    ObjectStep(name, fb.result())
-  }
-}
-
-final private class SumSchema[R, A](
-  _members: => List[(String, List[Any], Schema[R, Any])],
-  info: TypeInfo,
-  annotations: List[Any]
-)(ordinal: A => Int)
-    extends Schema[R, A] {
-  private lazy val members = _members
-
-  private lazy val subTypes = members.map { (label, subTypeAnnotations, schema) =>
-    (label, schema.toType_(), subTypeAnnotations)
-  }.sortBy(_._1)
-
-  private lazy val schemas = members.map(_._3).toVector // Vector's .apply is O(1) vs List's O(N)
-
-  private lazy val isEnum = subTypes.forall { (_, t, _) =>
-    t.allFields.isEmpty && t.allInputFields.isEmpty
-  }
-
-  private val isInterface = annotations.contains(GQLInterface())
-  private val isUnion     = annotations.contains(GQLUnion())
-
-  def toType(isInput: Boolean, isSubscription: Boolean): __Type =
-    if (!isInterface && !isUnion && subTypes.nonEmpty && isEnum) mkEnum(annotations, info, subTypes)
-    else if (!isInterface)
-      makeUnion(
-        Some(getName(annotations, info)),
-        getDescription(annotations),
-        subTypes.map(_._2).distinctBy(_.name).map(fixEmptyUnionObject),
-        Some(info.full),
-        Some(getDirectives(annotations))
-      )
-    else {
-      val impl = subTypes.map(_._2.copy(interfaces = () => Some(List(toType(isInput, isSubscription)))))
-      mkInterface(annotations, info, impl)
-    }
-
-  def resolve(value: A): Step[R] = schemas(ordinal(value)).resolve(value)
-}
 
 private object DerivationUtils {
